requirements.txt
flask
pandas
google-generativeai
python-dotenv
Pillow
